make organic editor state manage its own camera
